refactor(openomni): drop dead ingress pre-run path, merge middleware
The unrouted runPreRun chain (blacklist + channel-grant checks) had zero
production callers; blacklist and channel-grant enforcement is owned by
the routing pipeline (routing-runtime + resolve-route), pinned by
test/ingress/blacklist.test.ts and channel-grant.test.ts. Middleware
folds from eight files into ingress-authority.ts plus the actor helpers,
keeping only the routed pre-run surface (schema/coordinator/authority/mode).

Roadmap: #215 prep
