+++ packages/openomni/test/policy/ingress-authority-validation.test.ts
@@ -1,7 +1,9 @@
-import { afterEach, beforeEach, describe, expect, test } from "bun:test";
+import { describe, expect, test } from "bun:test";
 import type { Ingress, Policy } from "@openomni/protocol";
-import { ChannelGrantStore, Storage } from "@openomni/session";
-import { IngressAuthorityMiddleware } from "../../src/ingress/middleware/ingress-authority";
+import {
+  applyChannelGrantTreatment,
+  IngressAuthorityMiddleware,
+} from "../../src/ingress/middleware/ingress-authority";
 
 type RestrictedTrustTier = "observer" | "collaborator" | "assigned_worker";
 
@@ -28,20 +30,6 @@ const stubCoordinator = {
 };
 
 describe("IngressAuthorityMiddleware trust and validation", () => {
-  beforeEach(() => {
-    Storage.reset();
-    Storage.initialize({ dbPath: ":memory:" });
-    ChannelGrantStore.put({
-      id: "grant-test",
-      surface: "test",
-      kind: "trusted_channel",
-      createdBy: "act_owner",
-    });
-  });
-
-  afterEach(() => {
-    Storage.reset();
-  });
   test("canonical observer trust tier overrides legacy user role", async () => {
     const decisions: Policy.PolicyDecision[] = [];
     const event = makeInboundEvent({
@@ -49,7 +37,7 @@ describe("IngressAuthorityMiddleware trust and validation", () => {
     });
 
     await expect(
-      IngressAuthorityMiddleware.runPreRun({
+      IngressAuthorityMiddleware.runRoutedPreRun({
         event,
         coordinator: stubCoordinator,
         onDecision: (decision) => {
@@ -81,33 +69,55 @@ describe("IngressAuthorityMiddleware trust and validation", () => {
     });
 
     await expect(
-      IngressAuthorityMiddleware.runPreRun({
+      IngressAuthorityMiddleware.runRoutedPreRun({
         event,
         coordinator: stubCoordinator,
       }),
     ).rejects.toThrow("actor is not authorized to create top-level inbound work");
   });
 
-  test("uses channel default trust tier for unregistered actors", async () => {
-    ChannelGrantStore.put({
-      id: "grant-public-observer",
-      surface: "test",
-      channel: "public",
-      kind: "trusted_channel",
-      defaultTier: "observer",
-      createdBy: "act_owner",
-    });
+  test("channel grant treatment applies default trust tier to unregistered actors", () => {
     const event = makeInboundEvent({
       channel: "public",
       meta: { actor: { role: "user", id: "external-user-1" } },
     });
 
-    await expect(
-      IngressAuthorityMiddleware.runPreRun({
-        event,
-        coordinator: stubCoordinator,
-      }),
-    ).rejects.toThrow("actor is not authorized to create top-level inbound work");
+    const treated = applyChannelGrantTreatment(
+      event as Ingress.DirectEvent,
+      {
+        id: "grant-public-observer",
+        surface: "test",
+        channel: "public",
+        kind: "trusted_channel",
+        defaultTier: "observer",
+        createdBy: "act_owner",
+      },
+      "normal",
+    );
+
+    expect(treated.meta?.actor).toMatchObject({ role: "user", trustTier: "observer" });
+    expect(treated.meta?.channelGrantId).toBe("grant-public-observer");
+    expect(treated.meta?.inboundTreatment).toBe("normal");
+  });
+
+  test("channel grant treatment never overrides an explicit trust tier", () => {
+    const event = makeInboundEvent({
+      meta: { actor: { role: "user", actorId: "act_manager", trustTier: "manager" } },
+    });
+
+    const treated = applyChannelGrantTreatment(
+      event as Ingress.DirectEvent,
+      {
+        id: "grant-default-observer",
+        surface: "test",
+        kind: "trusted_channel",
+        defaultTier: "observer",
+        createdBy: "act_owner",
+      },
+      "normal",
+    );
+
+    expect(treated.meta?.actor).toMatchObject({ trustTier: "manager" });
   });
 
   test("allows explicit resident target when coordinator is missing", async () => {
@@ -116,7 +126,7 @@ describe("IngressAuthorityMiddleware trust and validation", () => {
       meta: { actor: { role: "user" } },
     });
 
-    const result = await IngressAuthorityMiddleware.runPreRun({ event });
+    const result = await IngressAuthorityMiddleware.runRoutedPreRun({ event });
 
     expect(result.target.kind).toBe("resident");
     expect(result.coordinator).toBeUndefined();
@@ -125,7 +135,7 @@ describe("IngressAuthorityMiddleware trust and validation", () => {
   test("aborts when coordinator is missing for worker creation target", async () => {
     const event = makeInboundEvent({ target: { kind: "worker" } });
 
-    await expect(IngressAuthorityMiddleware.runPreRun({ event })).rejects.toThrow(
+    await expect(IngressAuthorityMiddleware.runRoutedPreRun({ event })).rejects.toThrow(
       "coordinator is required for worker target",
     );
   });
@@ -134,7 +144,7 @@ describe("IngressAuthorityMiddleware trust and validation", () => {
     const badEvent = { not: "valid" };
 
     await expect(
-      IngressAuthorityMiddleware.runPreRun({
+      IngressAuthorityMiddleware.runRoutedPreRun({
         event: badEvent,
         coordinator: stubCoordinator,
       }),
@@ -146,7 +156,7 @@ describe("IngressAuthorityMiddleware trust and validation", () => {
     (event as Record<string, unknown>).mode = "fork";
 
     await expect(
-      IngressAuthorityMiddleware.runPreRun({
+      IngressAuthorityMiddleware.runRoutedPreRun({
         event,
         coordinator: stubCoordinator,
       }),
@@ -154,12 +164,12 @@ describe("IngressAuthorityMiddleware trust and validation", () => {
   });
 
   test("collects policy decisions via onDecision callback", async () => {
-    const decisions: unknown[] = [];
+    const decisions: Policy.PolicyDecision[] = [];
     const event = makeInboundEvent({
       meta: { actor: { role: "user" } },
     });
 
-    await IngressAuthorityMiddleware.runPreRun({
+    await IngressAuthorityMiddleware.runRoutedPreRun({
       event,
       coordinator: stubCoordinator,
       onDecision: (d) => {
@@ -174,20 +184,23 @@ describe("IngressAuthorityMiddleware trust and validation", () => {
     }
   });
 
-  test("registrations produce all six middleware steps", () => {
-    const state = {
-      input: makeInboundEvent(),
-      coordinator: stubCoordinator,
-    } satisfies Parameters<typeof IngressAuthorityMiddleware.registrations>[0];
-    const regs = IngressAuthorityMiddleware.registrations(state);
+  test("routed pre-run never re-runs blacklist or channel-grant checks", async () => {
+    const decisions: Policy.PolicyDecision[] = [];
+    const event = makeInboundEvent({
+      meta: { actor: { role: "user" } },
+    });
 
-    expect(regs).toHaveLength(6);
-    const names = regs.map((r) => r.name);
-    expect(names).toContain(IngressAuthorityMiddleware.CoordinatorPresence.name);
-    expect(names).toContain(IngressAuthorityMiddleware.SchemaValidation.name);
-    expect(names).toContain(IngressAuthorityMiddleware.BlacklistCheck.name);
-    expect(names).toContain(IngressAuthorityMiddleware.ChannelGrantCheck.name);
-    expect(names).toContain(IngressAuthorityMiddleware.AuthorityCheck.name);
-    expect(names).toContain(IngressAuthorityMiddleware.ModeDispatch.name);
+    await IngressAuthorityMiddleware.runRoutedPreRun({
+      event,
+      coordinator: stubCoordinator,
+      onDecision: (d) => {
+        decisions.push(d);
+      },
+    });
+
+    const policyIds = decisions.map((d) => d.policyId);
+    expect(policyIds).toContain("ingress.schema");
+    expect(policyIds).not.toContain("ingress.blacklist");
+    expect(policyIds).not.toContain("ingress.channel_grant");
   });
 });
